+++ CHANGELOG.md
@@ -7,9 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- XBlock SDK Workbench scenarios to simplify development.
+
 ### Changed
 
 - Update bundled VideoJS to 5.13.1.
+
+## [0.6.6] - 2017-05-03
+
+### Changed
+
+- Make transcripts help hint more helpful.
 
 ## [0.6.5] - 2017-04-21
 
@@ -171,4 +181,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [0.6.3]: https://github.com/raccoongang/xblock-video/compare/v0.6.2...v0.6.3
 [0.6.4]: https://github.com/raccoongang/xblock-video/compare/v0.6.3...v0.6.4
 [0.6.5]: https://github.com/raccoongang/xblock-video/compare/v0.6.4...v0.6.5
-[Unreleased]: https://github.com/raccoongang/xblock-video/compare/v0.6.5...HEAD
+[0.6.6]: https://github.com/raccoongang/xblock-video/compare/v0.6.5...v0.6.6
+[Unreleased]: https://github.com/raccoongang/xblock-video/compare/v0.6.6...HEAD